Abstract away common functionality of Evald
